requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydicom==2.4.4
nibabel==5.2.1
numpy==1.26.4
Pillow==10.4.0
python-multipart==0.0.12
